fix: resolve type conversion errors (double to int) for database fields
- Fixed restore_fix_service.dart: proper int/double handling for debts and ledger
- Fixed google_drive_delta_sync.dart: amount field now uses _asInt with default
- Fixed salary_entitlement_service.dart: proper type conversions for expense.amount
- Fixed salary_entitlements_screen.dart: basicSalary converted to int

All database amount fields are now IntColumn, with proper type conversions

+++ mobile/lib/screens/employees/salary_entitlements_screen.dart
@@ -826,7 +826,7 @@ class _SalaryEntitlementsScreenState
     await (db.update(db.employees)..where((t) => t.id.equals(ent.employee.id)))
         .write(
       EmployeesCompanion(
-        basicSalary: Value(newSalary),
+        basicSalary: Value(newSalary.toInt()),  // ⭐ int
         name: Value(nameCtrl.text.trim()),
       ),
     );
